test: Add unit tests for in-mem UniqueMultimap
